Changed max sig ID value from 32 bit to 64 bit

+++ ISSLScoreOfftargetsMMF/ISSLScoreOfftargetsMMF.cpp
@@ -20,7 +20,7 @@ bool calcMit = false;
 double maximum_sum;
 
 struct OverflowEntry {
-    uint32_t signatureId;
+    uint64_t signatureId;
     uint32_t occurrences;
 };
